+++ CIMS_Server 2.0/src/Field_Operations/DAO/VehicleDAO.java
@@ -45,7 +45,7 @@ public interface VehicleDAO {
      * @param id 
      * @return Vehicle with @param id
      */
-    public Vehicle find(int id);
+    public Vehicle find(long id);
 
     /**
      * Find all Vehicles in the database.
