+++ CIMS_Server 2.0/src/Field_Operations/Domain/Material.java
@@ -28,7 +28,7 @@ public class Material implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String name;
     private int type;
 
@@ -38,9 +38,9 @@ public class Material implements Serializable {
     /**
      * gets id of this material
      *
-     * @return int with id
+     * @return long with id
      */
-    public int getId() {
+    public long getId() {
         return id;
     }
